remove legacy assembly register splitter
This finally removes the legacy register splitting algorithm altogether.
