+++ pkg/schema/agnostic/assignment.go
@@ -19,8 +19,10 @@ import (
 
 	"github.com/consensys/go-corset/pkg/schema/register"
 	"github.com/consensys/go-corset/pkg/util/collection/array"
-	"github.com/consensys/go-corset/pkg/util/collection/stack"
+	"github.com/consensys/go-corset/pkg/util/collection/bit"
+	"github.com/consensys/go-corset/pkg/util/field"
 	"github.com/consensys/go-corset/pkg/util/math"
+	util_math "github.com/consensys/go-corset/pkg/util/math"
 	"github.com/consensys/go-corset/pkg/util/poly"
 )
 
@@ -60,9 +62,7 @@ func (p *Assignment) String(env register.Map) string {
 	}
 	//
 	builder.WriteString(" := ")
-	builder.WriteString(poly.String(p.RightHandSide, func(id register.Id) string {
-		return env.Register(id).Name
-	}))
+	builder.WriteString(StaticPoly2String(p.RightHandSide, env))
 	//
 	builder.WriteString(fmt.Sprintf("]^%d", p.Width(env)))
 	//
@@ -82,108 +82,6 @@ func (p *Assignment) Width(env register.Map) uint {
 	)
 	//
 	return max(lhs, rhs)
-}
-
-// Overflow calculates the amount of overflow in this assignment.  Here,
-// overflow is the number of additional bits of information generated by the
-// left-hand side compared with that consumed by the right-hand side.
-func (p *Assignment) Overflow(env register.Map) (uint, uint) {
-	var (
-		// Determine lhs width
-		lhs = CombinedWidthOfRegisters(env, p.LeftHandSide...)
-		// Determine rhs width
-		pos, neg = SplitWidthOfPolynomial(p.RightHandSide, StaticEnvironment(env))
-	)
-	// Normalise positive contribution
-	if pos > lhs {
-		pos = pos - lhs
-	} else {
-		pos = 0
-	}
-	// Normalise negative contribution
-	if neg > lhs {
-		neg = neg - lhs
-	} else {
-		neg = 0
-	}
-	//
-	return pos, neg
-}
-
-// Terminate performs some sanity checks on this
-func (p *Assignment) Terminate(env register.Map) (Assignment, StaticPolynomial) {
-	var (
-		tmp StaticPolynomial
-		// Determine lhs width
-		lhs = CombinedWidthOfRegisters(env, p.LeftHandSide...)
-		// Determine rhs width
-		rhs, _ = WidthOfPolynomial(p.RightHandSide, StaticEnvironment(env))
-	)
-	// Sanity check this assignment is really terminating.  That means all bits
-	// on the right-hand side must be consumed by bits on the left-hand side.
-	if lhs < rhs {
-		panic(fmt.Sprintf("unaccounted for carry bits (%dbits := %dbits)", lhs, rhs))
-	}
-	// Return without change
-	return *p, tmp.Set()
-}
-
-// Link attempts to resolve, for a non-terminating internal assignment, any
-// imbalance between the number of bits produced by the right-hand side of an
-// assignment, and the number consumed by the left-hand side.  If the more bits
-// are produced than are consumed, then a carry register is introduced to make
-// up the difference.
-func (p *Assignment) Link(env RegisterAllocator) (Assignment, StaticPolynomial) {
-	var overflow, underflow = p.Overflow(env)
-	// Check whether anything to resolve
-	if p.RightHandSide.Signed() {
-		return p.linkSigned(max(overflow, underflow), env)
-	} else if underflow > 0 {
-		// Unsigned assignments should not have underflow!
-		panic("unreachable")
-	}
-	//
-	return p.linkUnsigned(overflow, env)
-}
-
-func (p *Assignment) linkUnsigned(overflow uint, env RegisterAllocator) (Assignment, StaticPolynomial) {
-	var carry StaticPolynomial
-	// Sanity check whether actually need to do anything.
-	if overflow == 0 {
-		return *p, carry.Set()
-	}
-	// Allocate new register to get an Id
-	carryRegId := env.Allocate("c", overflow)
-	// Construct carry to be propagated forward
-	carry = carry.Set(poly.NewMonomial(one, carryRegId))
-	// Update left-hand side to include carry
-	lhs := array.Append(p.LeftHandSide, carryRegId)
-	// Resolved
-	return NewAssignment(lhs, p.RightHandSide), carry
-}
-
-func (p *Assignment) linkSigned(underflow uint, env RegisterAllocator) (Assignment, StaticPolynomial) {
-	var (
-		tmp     StaticPolynomial
-		carry   []StaticMonomial
-		lhs     = p.LeftHandSide
-		signBit = env.Allocate("s", 1)
-	)
-	// Append sign bit to borrow calculation
-	carry = append(carry, poly.NewMonomial(*math.NegPow2(underflow), signBit))
-	// Add carry component (if applicable)
-	if underflow > 0 {
-		// Allocate new register to get an Id
-		carryRegId := env.Allocate("c", underflow)
-		// Construct carry to be propagated forward
-		carry = append(carry, poly.NewMonomial(one, carryRegId))
-		// Update left-hand side to include carry
-		lhs = array.Append(lhs, carryRegId)
-	}
-	// Update left-hand side to include sign bit
-	lhs = array.Append(lhs, signBit)
-	// Resolved
-	return NewAssignment(lhs, p.RightHandSide), tmp.Set(carry...)
 }
 
 // Split an assignment according to a given field bandwidth.  This creates one
@@ -234,303 +132,500 @@ func (p *Assignment) linkSigned(underflow uint, env RegisterAllocator) (Assignme
 // For (sub-)assignments which still exceed the bandwidth requirement (such as
 // above), we must further split them by introducing additional temporary
 // registers.
-func (p *Assignment) Split(bandwidth uint, env RegisterAllocator) []Assignment {
-	var (
-		// worklist of remaining assignments
-		worklist stack.Stack[Assignment]
-		// set of completed assignments
-		completed []Assignment
-	)
-	// Initialise worklist
-	worklist.Push(*p)
-	// Continue splitting until no assignments outstanding.
-	for !worklist.IsEmpty() {
-		next := worklist.Pop()
-		// further splitting required?
-		if next.Width(env) > bandwidth {
-			// yes
-			worklist.PushReversed(next.innerSplit(bandwidth, env))
-		} else {
-			// no
-			completed = append(completed, next)
-		}
+func (p *Assignment) Split(field field.Config, env RegisterAllocator) (eqs []Assignment) {
+	// Check whether any splitting required
+	if p.Width(env) > field.BandWidth {
+		// Yes!
+		eqs = p.chunkUp(field, env)
+	} else {
+		// Nope
+		eqs = []Assignment{*p}
 	}
-	// Done
-	return completed
+	//
+	return eqs
 }
 
-// innerSplit performs one split of the given assignment according to the given
-// bandwidth, but does not guarantee that the resulting assignments fit within
-// the given bandwidth.  Rather, the resulting assignments may themselves need
-// to be split further.
-func (p *Assignment) innerSplit(bandwidth uint, env RegisterAllocator) []Assignment {
-	var assignments []Assignment = p.initialiseSplit(env)
-	// Merge to exploit available bandwidth.
-	assignments = coalesceAssignments(assignments, bandwidth, env)
+// Cap all terms within a polynomial to ensure they can be safely evaluated
+// within the given bandwidth.  For example, consider the following constraint
+// (where both registers are u8):
+//
+// 0 == X * Y
+//
+// Suppose a bandwidth of 15bits.  Then, X*Y cannot be safely evaluated since it
+// requires 16bits of information.  Instead, we have to break up either X or Y
+// into smaller chunks.  Suppose we break X into two 4bit chunks, X'0 and X'1.
+// Then we have:
+//
+// 0 == (256*X'1 + X'0) * Y
+//
+// --> 0 == 16*X'1*Y + X'0*Y
+//
+// At this point, each term can be safely evaluated within the given bandwidth
+// and this equation can be chunked.  Observe that we assume supplementary
+// constraints are included to enforce that X == 16*X'1 + X'0.
+//
+// The real challenge with this algorithm is, for a polynomial which cannot be
+// chunked, to determine which variable(s) to subdivide and by how much.
+func (p *Assignment) chunkUp(field field.Config, mapping RegisterAllocator) []Assignment {
+	var (
+		last      StaticPolynomial
+		iteration = 0
+		// Record initial number of registers
+		n = uint(len(mapping.Registers()))
+		//
+		splitter = NewRegisterSplitter(n)
+		// Determine the bitwidth of each chunk
+		rhsChunks []RhsChunk
+		lhsChunks []LhsChunk
+		//
+		initLhsChunks = initialiseLhsChunks(p.LeftHandSide, field, mapping)
+	)
+	//
+	for {
+		var (
+			overflows bit.Set
+			// Right-hand side
+			right = splitter.Apply(p.RightHandSide, mapping)
+		)
+		// Attempt to chunk right-hand side
+		lhsChunks, rhsChunks, overflows = determineRhsChunks(right, initLhsChunks, field, mapping)
+		//
+		if overflows.Count() == 0 {
+			// Successful chunking, therefore include any constraints necessary
+			// for splitting of non-linear terms and construct final equations.
+			break
+		} else if iteration > 1 && right.Equal(last) {
+			// If we get here, then splitting made no improvement on the
+			// previous iteration and we are stuck.  The idea is that this
+			// should be unreachable, but it remains as a fall-back to prevent
+			// the potential for an infinite loop.
+			debugChunks(lhsChunks, rhsChunks, mapping)
+			fmt.Printf("Divisions: %s\n", splitter.String(mapping))
+			panic(fmt.Sprintf("malformed assignment (after %d iterations)", iteration))
+		}
+		// Update divisions based on identified overflows
+		splitter.Subdivide(overflows)
+		// Reset any allocated carry registers as we are starting over
+		mapping.Reset(n)
+		splitter.Reset()
+		// Start next iteration
+		last = right
+		iteration++
+	}
+	// Initialise with splits
+	assignments := splitter.assignments
+	// Reconstruct equations
+	for i := range len(lhsChunks) {
+		l := lhsChunks[i]
+		r := rhsChunks[i]
+		//
+		assignments = append(assignments, NewAssignment(l.contents, r.contents))
+	}
 	// Done
 	return assignments
 }
 
-// InitialiseSplit constructs an initial set of assignments with one for each
-// target register.  Here, each assignment includes all monomials from the
-// right-hand side whose coefficient begins within the range of the
-// corresponding target register.  For example, consider the following
-// assignment:
-//
-// b, y'1, y'0 = 2^8*x'1 + x'0 + 1
-//
-// would be initialised as follows:
-//
-// y'0 := x'0 + 1 ; y'1 := x'1 ; b := 0
-//
-// Observe that b has no initial assignment.  This is because, in the end, it
-// will form part of the previous assignment (i.e. after grouping).
-//
-// One mildly complicating factor is that of "dividing monomials".  Consider our
-// example, above where the term 2^8*x'1 becomes just x'1.  This makes sense as,
-// in the final assignment "y'1 := x'1", we know that y'1 start at bit offset 8.
-// Thus, we can see that "y'1 := 2^8*x'1" doesn't make sense.  Thus, to
-// determine the right coefficients, a division process is employed.  Since
-// division may not be exact, remainders are left to be processed again and a
-// worklist is used to manage those bits that still need processing.
-func (p *Assignment) initialiseSplit(env register.Map) []Assignment {
+func initialiseLhsChunks(regs []register.Id, field field.Config, mapping register.Map) []LhsChunk {
 	var (
-		// Final list of assignments to be constructed
-		monomials = make([][]StaticMonomial, len(p.LeftHandSide))
-		// Worklist contains list of monomials being processed.
-		worklist stack.Stack[StaticMonomial]
-		// Assignments to be constructed
-		assignments = make([]Assignment, len(p.LeftHandSide))
+		chunks     []Chunk[[]register.Id]
+		chunkWidth = determineInitialChunkWidth(field)
 	)
-	// Initialise worklist from source polynomial
-	for j := range p.RightHandSide.Len() {
-		worklist.Push(p.RightHandSide.Term(j))
+	//
+	for len(regs) != 0 {
+		var chunk Chunk[[]register.Id]
+		// Determine next chunkd
+		chunk, regs = getNextLhsChunk(regs, chunkWidth, mapping)
+		chunks = append(chunks, chunk)
 	}
-	// Continue processing monomials until no more remain.
-	for !worklist.IsEmpty() {
-		// Extract next item to process
-		next := worklist.Pop()
-		// Identify target register
-		i, offset := identifyEnclosingRegister(p.LeftHandSide, next.Coefficient(), env)
-		// Divide monomial by bit offset
-		next, rest := next.Shr(offset)
-		// Check wether division was exact
-		if !rest.IsZero() {
-			// No, therefore some remainder must still be processed
-			worklist.Push(rest)
+	//
+	return chunks
+}
+
+// Determining the chunkwidth to use for initialising the left-hand side is
+// somewhat subtle, and can impact both the performance of the algorithm and the
+// overall chance of success.  In particular, it is useful to have at least one
+// additional bit over the field's register bitwidth to account for sign bits.
+// The chosen solution here is a heuristic which aims to ensure: (i) there is at
+// least one additional bit of information; (ii) there are unused bits in the
+// given bandwidth (e.g. as needed for carries, etc).
+func determineInitialChunkWidth(field field.Config) uint {
+	//var delta = max(1, (field.BandWidth-field.RegisterWidth)/4)
+	var delta = uint(1)
+	//
+	return field.RegisterWidth + delta
+}
+
+func getNextLhsChunk(regs []register.Id, chunkWidth uint, mapping register.Map) (LhsChunk, []register.Id) {
+	var bitwidth uint
+	//
+	for i, r := range regs {
+		reg := mapping.Register(r)
+		//
+		if bitwidth+reg.Width > chunkWidth {
+			return LhsChunk{bitwidth, regs[:i]}, regs[i:]
 		}
 		//
-		monomials[i] = append(monomials[i], next)
+		bitwidth += reg.Width
 	}
-	// Finally construct assignments
-	for i, lid := range p.LeftHandSide {
-		var tmp StaticPolynomial
-		// Construct ith assignment
-		assignments[i] = Assignment{
-			// left-hand side
-			[]register.Id{lid},
-			// right-hand side
-			tmp.Set(monomials[i]...),
+	//
+	return LhsChunk{bitwidth, regs}, nil
+}
+
+// Divide a polynomial into "chunks", each of which has a maximum bitwidth as
+// determined by the chunk widths.  This inserts carry and borrow registers as
+// needed to ensure correctness of both signed and unsigned arithmetic.
+func determineRhsChunks(p StaticPolynomial, chunks []LhsChunk, field field.Config,
+	mapping RegisterAllocator) ([]LhsChunk, []RhsChunk, bit.Set) {
+	//
+	var (
+		env       = StaticEnvironment(mapping)
+		rhsChunks []RhsChunk
+		lhsChunks []LhsChunk
+		vars      bit.Set
+		signed    bool
+	)
+	// Subdivide polynomial into chunks
+	for i, chunk := range chunks {
+		var (
+			last      = i+1 == len(chunks)
+			remainder StaticPolynomial
+		)
+		// Chunk the polynomial
+		p, remainder = p.Shr(chunk.bitwidth)
+		// Determine chunk width
+		chunkWidth, s := RawWidthOfPolynomial(remainder, env)
+		// Determine width of overflow
+		overflow := chunkWidth - chunk.bitwidth
+		// Check whether signed arithmetic begins
+		signed = signed || s
+		// Check whether chunk fits
+		if chunkWidth > field.BandWidth {
+			// No, it does not.
+			vars.Union(RegisterReadSet(remainder))
+		} else if !last && chunkWidth > chunk.bitwidth {
+			// Overflow case.
+			p, chunk = propagateCarry(chunk, overflow, p, mapping)
 		}
+		// Manage signed arithmetic
+		if signed && !last {
+			// Overflow case.
+			p, chunk = propagateBorrow(chunk, overflow, p, mapping)
+		}
+		//
+		rhsChunks = append(rhsChunks, RhsChunk{chunkWidth, remainder})
+		lhsChunks = append(lhsChunks, chunk)
+	}
+	//
+	return lhsChunks, rhsChunks, vars
+}
+
+// Propagate a given number of overflow bits from the current chunk into the
+// polynomial being carried forward into the next chunk.  For example, consider
+// these two chunks (viewed as instructions):
+//
+// var x'0, x'1, y'0, y'1 u8
+// var c u2
+//
+//    x'0 = 3 * y'0
+// c, x'1 = y'1
+//
+// Looking at the first statement, we have the following alignment of bits:
+//
+//         9 8 7 6 5 4 3 2 1 0
+//            +-+-+-+-+-+-+-+-+
+// x'0        | | | | | | | | |
+//        +-+-+-+-+-+-+-+-+-+-+
+// 3*y'0: | | | | | | | | | | |
+//        +-+-+-+-+-+-+-+-+-+-+
+//
+// We can see the right-hand side has an overflow of 2 bits.  Thus, we need to
+// allocate a "carry register" of the given size and then propagate this into
+// the next chunk (i.e. instruction).  That gives the following:
+//
+// var x'0, x'1, y'0, y'1 u8
+// var c, c$2 u2
+//
+// c$2, x'0 = 3 * y'0
+//   c, x'1 = y'1 + c$2
+//
+// Here, c$2 is the carry register allocated to balance the first assignment,
+// and this then must be propagated into the second instruction.
+
+func propagateCarry(chunk LhsChunk, overflow uint, carry StaticPolynomial,
+	mapping RegisterAllocator) (StaticPolynomial, LhsChunk) {
+	// Overflow case.
+	var tmp StaticPolynomial
+	// Allocate new register to get an Id
+	carryRegId := mapping.Allocate("c", overflow)
+	// Propage carry forward
+	carry = carry.Add(tmp.Set(poly.NewMonomial(one, carryRegId)))
+	// include carry in lhs
+	return carry, LhsChunk{chunk.bitwidth, array.Append(chunk.contents, carryRegId)}
+}
+
+// Propagate a borrow bit from the current chunk into the polynomial being
+// carried forward into the next chunk.  For example, consider these two chunks
+// (viewed as instructions):
+//
+// var x'0, x'1, y'0, y'1 u8
+// var b u1
+//
+//    x'0 = y'0 - 1
+// b, x'1 = y'1
+//
+// In this case, we have 9bits of information from the right-hand side of the
+// first instruction flowing into only 8bits on the left-hand side.  To make
+// this work, we must introduce a specific "sign bit" to propagate the borrow
+// originating in the first instruction into the second, like so:
+//
+// var x'0, x'1, y'0, y'1 u8
+// var b, b$2 u1
+//
+// b$2, x'0 = y'0 - 1
+//   b, x'1 = y'1 - b$2
+//
+// Here, b$2 is the allocated sign bit to account for the potential underflow in
+// the first instruction.
+
+func propagateBorrow(chunk LhsChunk, overflow uint, carry StaticPolynomial,
+	mapping RegisterAllocator) (StaticPolynomial, LhsChunk) {
+	var borrow StaticPolynomial
+	// Allocate new register to get an Id
+	signBit := mapping.Allocate("b", 1)
+	// Put sign bit after carry (i.e. overflow)
+	carry = carry.Add(borrow.Set(poly.NewMonomial(*math.NegPow2(overflow), signBit)))
+	// include sign in lhs
+	return carry, LhsChunk{chunk.bitwidth, array.Append(chunk.contents, signBit)}
+}
+
+// Chunk represents a "chunk of information bits".
+type Chunk[T any] struct {
+	bitwidth uint
+	contents T
+}
+
+// LhsChunk captures the chunk type used for the Left-Hand Side (LHS) of an
+// assignment.
+type LhsChunk = Chunk[[]register.Id]
+
+// RhsChunk captures the chunk type used for the Right-Hand Side (RHS) of an
+// assignment.
+type RhsChunk = Chunk[StaticPolynomial]
+
+// StaticPoly2String provides a convenient helper function for debugging polynomials.
+func StaticPoly2String(p StaticPolynomial, env register.Map) string {
+	return poly.String(p, func(r register.Id) string {
+		return env.Register(r.Id()).Name
+	})
+}
+
+// RegisterSplitter is used to manage the mechanism of splitting variables into
+// limbs in order to improve the precision of a chunk.
+type RegisterSplitter struct {
+	divisions   []uint
+	assignments []Assignment
+	parents     map[register.Id]register.Id
+}
+
+// NewRegisterSplitter constructs a new splitter for a given number of variables.
+func NewRegisterSplitter(n uint) RegisterSplitter {
+	var (
+		divisions = make([]uint, n)
+		parents   = make(map[register.Id]register.Id)
+	)
+	//
+	for i := range divisions {
+		divisions[i] = 1
+	}
+	//
+	return RegisterSplitter{divisions, nil, parents}
+}
+
+// ParentOf gets the original source variable from which this variable was
+// derived (if it was indeed derived).  For example, if a register X is split
+// into limbs X'1 and X'0, then the parent of the two limbs is X.  In the case
+// that a register was not derived through splitting, then it is its own parent.
+func (p *RegisterSplitter) ParentOf(v register.Id) register.Id {
+	if parent, ok := p.parents[v]; ok {
+		return parent
+	}
+	//
+	return v
+}
+
+// Subdivide takes all registers in the given set and further subdivides them.
+// For example, if they were previously being divided in 2, then they will now
+// be divided in 4, etc.
+func (p *RegisterSplitter) Subdivide(vars bit.Set) bool {
+	var (
+		changed = false
+		parents bit.Set
+	)
+	// Normalise variables
+	for iter := vars.Iter(); iter.HasNext(); {
+		rid := register.NewId(iter.Next())
+		parents.Insert(p.ParentOf(rid).Unwrap())
+	}
+	// Update division
+	for i := range p.divisions {
+		if parents.Contains(uint(i)) {
+			p.divisions[i] *= 2
+			changed = true
+		}
+	}
+	//
+	return changed
+}
+
+// Reset assignments created for the current splitting, and the ledger of
+// parents.
+func (p *RegisterSplitter) Reset() {
+	p.assignments = nil
+	// Reset parent of relationship
+	p.parents = make(map[register.Id]register.Id)
+}
+
+// Apply the current subdivisions to a given polynomial.  Specifically, this
+// splits all registers into their limbs and subsitutes them into the
+// polynomial.  As a by-product this also records the assignments needed for the
+// mapping from registers to their limbs.
+func (p *RegisterSplitter) Apply(poly StaticPolynomial, mapping RegisterAllocator) StaticPolynomial {
+	var cache = make(map[register.Id]StaticPolynomial)
+	//
+	for i, div := range p.divisions {
+		var (
+			rid = register.NewId(uint(i))
+			reg = mapping.Register(rid)
+		)
+		//
+		if div != 1 && reg.Width > 1 {
+			var limbPoly = p.splitVariable(rid, div, mapping)
+			//
+			cache[rid] = limbPoly
+		}
+	}
+	//
+	return SubstitutePolynomial(poly, func(reg register.Id) StaticPolynomial {
+		if rp, ok := cache[reg]; ok {
+			return rp
+		}
+		// no substitution
+		return nil
+	})
+}
+
+func (p *RegisterSplitter) splitVariable(rid register.Id, div uint, mapping RegisterAllocator) (r StaticPolynomial) {
+	var (
+		one      big.Int = *big.NewInt(1)
+		rhs      StaticPolynomial
+		terms    []StaticMonomial
+		reg      = mapping.Register(rid)
+		maxWidth = reg.Width / div
+		bitwidth = reg.Width
+		width    uint
+	)
+	// Round up (if necessary)
+	if (maxWidth * div) < reg.Width {
+		maxWidth++
+	}
+	// Determine limb widths
+	limbWidths := register.LimbWidths(maxWidth, reg.Width)
+	// Allocate limbs
+	limbs := p.alloc(rid, limbWidths, mapping)
+	// Construct limb polynomial
+	for i, limb := range limbs {
+		var (
+			c         = util_math.Pow2(width)
+			limbWidth = min(bitwidth, limbWidths[i])
+		)
+		//
+		if limbWidth > 0 {
+			terms = append(terms, poly.NewMonomial(*c, limb))
+			width += limbWidth
+		}
+		//
+		bitwidth -= limbWidth
+	}
+	// Update assignments
+	assignment := NewAssignment(limbs, rhs.Set(poly.NewMonomial(one, rid)))
+	p.assignments = append(p.assignments, assignment)
+	//
+	return r.Set(terms...)
+}
+
+func (p *RegisterSplitter) String(mapping register.Map) string {
+	var (
+		builder strings.Builder
+		first   = true
+	)
+	//
+	builder.WriteString("[")
+	//
+	for i, div := range p.divisions {
+		rid := register.NewId(uint(i))
+		//
+		if div > 1 {
+			if !first {
+				builder.WriteString(", ")
+			}
+			//
+			name := mapping.Register(rid).Name
+			builder.WriteString(fmt.Sprintf("%s/%d", name, div))
+			//
+			first = false
+		}
+	}
+	//
+	builder.WriteString("]")
+	//
+	return builder.String()
+}
+
+// Allocate the limbs for a parent variable which is being split.  This records
+// the fact that those limbs were derived from this parent.
+func (p *RegisterSplitter) alloc(parent register.Id, limbWidths []uint, mapping RegisterAllocator) []register.Id {
+	var (
+		reg   = mapping.Register(parent)
+		limbs = mapping.AllocateN(reg.Name, limbWidths)
+	)
+	// Record parent of each limbs
+	for _, limb := range limbs {
+		p.parents[limb] = parent
 	}
 	// Done
-	return assignments
+	return limbs
 }
 
-// Identify enclosing register determines the index into the given regs array of
-// the register whose bitrange encloses the given value.  This also returns the
-// starting (bit) offset of this register.
-func identifyEnclosingRegister(regs []register.Id, value big.Int, env register.Map) (uint, uint) {
-	var bitOffset uint
-	//
-	for i, lid := range regs {
-		var reg = env.Register(lid)
-		// Value contained by this register?
-		if withinBitRange(value, bitOffset, bitOffset+reg.Width) {
-			// Yes!
-			return uint(i), bitOffset
-		}
-		// Shift offset along
-		bitOffset += reg.Width
-	}
-	// It should not be possible to get here if the original assignments were
-	// well-formed.  Right?
-	panic("unreachable")
-}
-
-// CoalesceAssignments attempts to merge consecutive assignments to exploit the
-// available bandwidth as much as possible.  For example, consider the example
-// initial splitting obtained above:
+// useful for debugging the splitting algorithm.
 //
-// [y'0 := x'0 + 1]^9 ; [y'1 := x'1]^8 ; [b := 0]^1
-//
-// Here, the width of each assignment is given as a superscript.  Roughly
-// speaking, any two assignments can be safely merged when the combined width
-// remains within the target bandwidth.  For our example, we merge the last two
-// assignments as follows:
-//
-// [y'0 := x'0 + 1]^9 ; [b, y'1 := x'1]^9
-//
-// This is permitted because the combined assignment still meets the necessary
-// bandwidth requirements.
-//
-// The process of combining assignments is not as simple as outlined above. This
-// is due to the need for carry registers which, at this point, have not yet
-// been added. Carry registers connect "overflowing" assignments together through
-// so-called carry registers.  Such registers are inspired by the carry bits
-// typically found in FLAGS' registers of CPU architectures.  One obvious
-// difference, however, is that they are not necessarily bits --- rather they
-// can have arbitrary
-//
-// An assignment is said to be overflowing when its right-hand side produces
-// more bits of information that its left-hand side consumes.  For example, this
-// assignment:
-//
-//	b, x'1, x'0 := 2^8*y'1 + y'0 + 1
-//
-// is split into two assignments as follows (assuming u8 limbs for x,y and u1
-// for b):
-//
-//	b, x'1 := y'1 ; x'0 := y'0 + 1
-//
-// Here, the assignment to x'0 is said to be overflowing because we have 9bits
-// of information on the right-hand side going into only 8bits on the left-hand
-// side.  As such, a u1 carry register (in this case) must be added to balance
-// the assignment which, in turn, must also be propagated up as follows:
-//
-//	b, x'1 := y'1 + c ; c, x'0 := y'0 + 1
-//
-// Observe that, whilst a 1bit carry was used in this case, the width of the
-// carry register is determined by the "overhang" --- that is, the difference
-// between the width of the right-hand side and the left-hand side.
-//
-// NOTE: to be clear, when the left-hand side is wider than the right-hand side,
-// no carry is required.
-//
-// NOTE: this implementation does not attempt to find an optimal allocation of
-// assignments (as this may indeed be a hard computational problem).  Instead,
-// assignments are merged greedily starting from the least significant position.
-func coalesceAssignments(assignments []Assignment, bandwidth uint, env RegisterAllocator) []Assignment {
-	var (
-		// Set of coalesced assignments
-		coalesced []Assignment
-		// Starting index of current sequence
-		start uint
-		// Width of current sequence being coalesced
-		width uint
-		// carry being propagated up
-		carry StaticPolynomial
-		// next assignment
-		next Assignment
-	)
-	//
-	for i, a := range assignments {
-		var nWidth uint = width + a.Width(env)
-		// FIXME: the nWidth calculation is incorrect in the presence of carry.
-		// Specifically, because any carry being propagated could increase the
-		// required width.
-		if nWidth > bandwidth {
-			next, carry = coalesce(false, assignments[start:i], carry, env)
-			// append coalesced group
-			coalesced = append(coalesced, next)
-			// reset for next group
-			start, width = uint(i), 0
-		} else {
-			width = nWidth
-		}
-	}
-	// Coalesce final group where carry can be ignored since it is already
-	// checked in Terminate.
-	next, _ = coalesce(true, assignments[start:], carry, env)
-	coalesced = append(coalesced, next)
-	//
-	return coalesced
-}
-
-// Coalesce a group of assignments together.  This is mostly straightforward,
-// but care must be taken to account for the bit offset of each assignment being
-// coalesced.  For example, consider coelescing these two assignments:
-//
-//	x'1 := 1 + y'1 ; x'0 := y'0
-//
-// The obvious (but incorrect) coalescing would be:
-//
-//	x'1, x'0 := 1 + y'1 + y'0
-//
-// Clearly, this is incorrect because y'1 is now overlapping with x'0 when
-// previously it overlapped with x'1.  Likewise, 1 is overlapping with x'0 not
-// x'1.  The correct coalescing (assuming u8 limbs) is:
-//
-//	x'1, x'0 := 256 + 2^8*y'1 + y'0
-//
-// The key difference is that an "offset" factor has been applied to the
-// monomial previously assigned to x'1.  The offset factor (in this case) is
-// determined by the width of x'0.
-func coalesce(last bool, assignments []Assignment, carry StaticPolynomial,
-	env RegisterAllocator) (Assignment, StaticPolynomial) {
-	//
-	var (
-		offset uint
-		lhs    []register.Id
-		rhs    StaticPolynomial
-	)
-	//
-	for i, ith := range assignments {
-		// update left-hand side
-		lhs = append(lhs, ith.LeftHandSide...)
-		// update right-hand side
-		if i == 0 {
-			rhs = ith.RightHandSide.Clone()
-			// Add carry (if applicable)
-			if carry != nil {
-				rhs = rhs.Add(carry)
-			}
-		} else {
-			var tmp = ith.RightHandSide
-			// apply offset factor
-			rhs = rhs.Add(tmp.MulScalar(math.Pow2(offset)))
-		}
-		// update bit offset
-		offset += CombinedWidthOfRegisters(env, ith.LeftHandSide...)
-	}
-	// Construct new assignment
-	assign := NewAssignment(lhs, rhs)
-	//
-	if last {
-		return assign.Terminate(env)
-	}
-	// Resolve overflow by allocating carry as required.
-	return assign.Link(env)
-}
-
-// withinBitRange checks whether a given integer value it contained within a
-// given bit range [s,e).  For example, 123 is contained within the range 0..8,
-// but 256 is not.
-func withinBitRange(val big.Int, start, end uint) bool {
-	var (
-		s = big.NewInt(2)
-		e = big.NewInt(2)
-	)
-	// Sanity check for zero needed as 2^0==1.
-	if start == 0 && val.Cmp(&zero) == 0 {
-		return true
-	}
-	//
-	s.Exp(s, big.NewInt(int64(start)), nil)
-	e.Exp(e, big.NewInt(int64(end)), nil)
-	// Check interval
-	return val.CmpAbs(s) >= 0 && val.CmpAbs(e) < 0
-}
-
 // nolint
-func debugAssignments(env RegisterAllocator, assignments ...Assignment) {
-	for i, ith := range assignments {
+func debugChunks(lhs []LhsChunk, rhs []RhsChunk, mapping register.Map) {
+	//
+	for i := len(lhs); i > 0; i-- {
+		ith := lhs[i-1]
+		fmt.Printf("[u%d ", ith.bitwidth)
 
-		if i != 0 {
-			fmt.Printf(" | ")
+		for j := len(ith.contents); j > 0; j-- {
+			rid := ith.contents[j-1]
+			if j < len(ith.contents) {
+				fmt.Printf(", ")
+			}
+
+			fmt.Print(mapping.Register(rid).Name)
 		}
-
-		fmt.Printf("%s", ith.String(env))
+		//
+		fmt.Print("]")
+	}
+	//
+	fmt.Print(" := ")
+	//
+	for i := len(rhs); i > 0; i-- {
+		ith := rhs[i-1]
+		fmt.Printf("[u%d ", ith.bitwidth)
+		//
+		fmt.Print(StaticPoly2String(ith.contents, mapping))
+		//
+		fmt.Print("]")
 	}
 	//
 	fmt.Println()
